Bad use of quint8
For whatever reason I had decided to use a quint8 type for some of the
parsing which in effect was causing bad parsing of some of the tags
since I should have probably been using quint32.  That is fixed, however
there does seem to be an issue with overrun -- sometimes the parser
comes close to the end of the tag and there's a bad frame size input.
Not sure why yet.