UI functionality
UI is much more usable and I can now use it to investigate some holes in
my ID3 parsing chopping characters off of tags.  I have a suspicion of
where it's at and it should hopefully be an easy fix.

I was pleasantly surprised at how easy it was to connect up the UI to
the SQL queries.  Compiled first try and preliminary testing seems to
indicate everything is working.

TODO - implement a way to select "All Artists" or "All Albums".  Once an
artist is selected there's no way to select all